refactor: Habit에 묶여있던 잔디 뷰 -> Grid로 분기처리
