+++ Habit Management/Feature/Grid/GridBackground/GridBackgroundView.swift
@@ -1,0 +1,25 @@
+//
+//  GridBackgroundView.swift
+//  Habit Management
+//
+//  Created by 서충원 on 6/12/25.
+//
+
+import SwiftUI
+
+struct GridBackgroundView<Content: View>: View {
+    private let content: Content
+    
+    init(@ViewBuilder content: () -> Content) {
+        self.content = content()
+    }
+    
+    var body: some View {
+        VStack {
+            content
+        }
+        .background(Color(hex: Color.HabitColor.primary.rawValue))
+        .clipShape(RoundedRectangle(cornerRadius: 12))
+        .scaledPadding(top: 0, leading: 16, bottom: 0, trailing: 16)
+    }
+}
