+++ Habit Management/Feature/Grid/GridFeature.swift
@@ -1,0 +1,28 @@
+//
+//  GridFeature.swift
+//  Habit Management
+//
+//  Created by 서충원 on 6/12/25.
+//
+
+import Foundation
+import ComposableArchitecture
+
+@Reducer
+struct GridFeature {
+    struct State: Equatable {
+        
+    }
+    
+    enum Action: Equatable {
+        
+    }
+    
+    var body: some ReducerOf<Self> {
+        Reduce { state, action in
+            switch action {
+                
+            }
+        }
+    }
+}
